Modified MipsBox with response queue and message queue.

+++ Mips-net/Module/ITwaveModule.cs
@@ -10,12 +10,12 @@ namespace Mips_net.Module
     public interface ITwaveModule
     {
         Task<double> GetTWaveFrequency(string channel);
-        Task<Unit> SetTWaveFrequency(string channel, int frequency);
+        Task<Unit> SetTWaveFrequency(string channel, double frequency);
         Task<double> GetTWavePulseVoltage(string channel);
-        Task<Unit> SetTWavePulseVoltage(string channel, int voltage);
-        Task<Unit> SetTWaveGuard1Voltage(string channel, int voltage);
+        Task<Unit> SetTWavePulseVoltage(string channel, double voltage);
+        Task<Unit> SetTWaveGuard1Voltage(string channel, double voltage);
         Task<double> GetTWaveGuard1Voltage(string channel);
-        Task<Unit> SetTWaveGuard2Voltage(string channel, int voltage);
+        Task<Unit> SetTWaveGuard2Voltage(string channel, double voltage);
         Task<double> GetTWaveGuard2Voltage(string channel);
         Task<BitArray> GetTWaveSequence(string channel);
         Task<Unit> SetTWaveSequence(string channel, BitArray sequence);
